% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{gsn_loglik}
\alias{gsn_loglik}
\title{Likelihood contribution to VB expected posterior}
\usage{
gsn_loglik(x_list, y_list, m, v, sigma)
}
\description{
Computes the term
 \sum_{r = 1}^{R} \sum_{i = 1}^{n_{r}} 1 / 2 * sigma ^ 2 ((y_{i}^{(r)} - x_{i}^{(r) T}m^{(r)})^2 + x_{i}^{(r) T} V^{(r)} x_{i}^{(r)}
}

