% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{mean_source_multinom}
\alias{mean_source_multinom}
\title{Compute deviation between variational means and latent sources}
\usage{
mean_source_multinom(n, pi, m, S, Psi_inv)
}
\description{
Computes the term
\sum_{r = 1}^{R}\sum_{k = 1}^{K} n_{r}\pi_{k}^{(r)} *
(m^{(r)} - s_{\cdot k})^{T}\Psi^{-1}(m^{(r)} - s_{\cdot k})
}

