% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{mulitnom_loglik}
\alias{mulitnom_loglik}
\title{Calculate expected multinomial likelihood term in posterior}
\usage{
mulitnom_loglik(n, pi, phi)
}
\description{
Computes
sum_{r = 1}^{R} \sum_{k = 1}^{K} pi_{k}^{(r)} \log \varphi_{k}
}

