% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{multinom_entropy}
\alias{multinom_entropy}
\title{Cross-Tasks multinomial entropy}
\usage{
multinom_entropy(pi)
}
\description{
Say r indexes tasks and k indexes latent source clusters.
Then, this term is 
sum_{r = 1}^{R} \sum_{k = 1}^{K} pi_{k}^{(r)} \log(\pi_{k}^{(r)})

I really should be using the log-sum-exp trick, but hopefully pi
never gets too small in these preliminary experiments.
}

