% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{sum_log_det}
\alias{sum_log_det}
\title{Compute sum of log determinants}
\usage{
sum_log_det(v)
}
\description{
\sum log|V^{(r)}|
Maybe it would be smarter to do a cholesky factorization sort of things.
}

