% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{update_m}
\alias{update_m}
\title{Update variational mean parameters}
\usage{
update_m(x, y, Psi_inv, S, pi, sigma)
}
\description{
\hat{m}^{(r)} &= \left(\frac{1}{\sigma^{2}}X^{(r) T}X^{(r)} + n_{r}\Psi^{-1}\right)^{-1}\left(\frac{1}{\sigma^{2}}X^{(r) T}y^{(r)} + n_{r}\Psi^{-1} S \pi^{(r)}\right)
}

