% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{update_pi}
\alias{update_pi}
\title{Update the variational clustering parameters}
\usage{
update_pi(m, S, Psi_inv, phi)
}
\description{
\hat{\pi}_{k}^{(r)} &\propto \varphi_{k}\exp{\left(m^{(r)} - s_{\cdot k}\right)^{T}\Psi^{-1}\left(m^{(r)} - s_{\cdot k}\right)}
}

