% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{update_Psi}
\alias{update_Psi}
\title{Update unshared covariance matrix Psi}
\usage{
update_Psi(v, m, S, pi)
}
\description{
\hat{Psi} &= \frac{1}{N}\sum_{r = 1}^{p_{1}}n_{r}\left(V^{(r)} + \sum_{k = 1}^{K}  \pi_{k}^{(r)}\left(m^{(r)} - s_{\cdot k}\right)\left(m^{(r)} - s_{\cdot k}\right)^{T}\right)
}

