% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{update_v}
\alias{update_v}
\title{Update a single V^{(r)}}
\usage{
update_v(x, Psi, sigma)
}
\description{
1 / n_{r}  * (\Psi^{-1} + \frac{1}{n_{r}\sigma^{2}}\sum_{i = 1}^{n_{r}}x_{i}^{(r)}x_{i}^{(r) T})^{-1}
}

