% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial.R
\name{vb_bound_multinom}
\alias{vb_bound_multinom}
\title{Multinomial Variational Lower Bound}
\usage{
vb_bound_multinom(data_list, var_list, param_list)
}
\description{
Compute the variational lower bound, equation (17) in the
reference, for the multinomial case.
}
\references{
Flexible Latent Variable Models for Multi-Task Learning
}

